Move all types to types folder
