+++ src/rules/SimpleRule.js
@@ -1,6 +1,5 @@
 /* @flow */
-import type {CSSStyleRule} from '../cssom'
-import type {RuleOptions, ToCssOptions, BaseRule} from '../types'
+import type {CSSStyleRule, RuleOptions, ToCssOptions, BaseRule} from '../types'
 
 export default class SimpleRule implements BaseRule {
   type = 'simple'
