+++ src/rules/StyleRule.js
@@ -2,8 +2,8 @@
 import warning from 'warning'
 import toCss from '../utils/toCss'
 import toCssValue from '../utils/toCssValue'
-import type { CSSStyleRule } from '../cssom';
 import type {
+  CSSStyleRule,
   ToCssOptions,
   RuleOptions,
   Renderer as RendererInterface,
